+++ lewp/src/module/mod.rs
@@ -12,3 +12,7 @@ pub type Modules = Vec<ModulePtr>;
 
 /// Wrapper type for a Module instance.
 pub type ModulePtr = Rc<RefCell<dyn Module>>;
+
+/// The unique module ID. Similar to the page ID, it is used to identify resources
+/// belonging to the module on the file system.
+pub type ModuleId = str;